requirements.txt
openai
graphdatascience
retry==0.9.2
langchain>=0.0.216
langchain-community
langchain-openai
langchain-core
streamlit==1.23.1
streamlit-chat==0.0.2.2
streamlit-chat-media==0.0.4
numpy
pandas
plotly==5.15.0
altair<5
python-dotenv==1.0.0